+++ adapters/mqtt-vertx-base/src/main/java/org/eclipse/hono/adapter/mqtt/AbstractVertxBasedMqttProtocolAdapter.java
@@ -20,6 +20,7 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.function.BiConsumer;
 import java.util.stream.Collectors;
 
@@ -33,7 +34,6 @@ import org.eclipse.hono.client.CommandResponse;
 import org.eclipse.hono.client.MessageConsumer;
 import org.eclipse.hono.client.MessageSender;
 import org.eclipse.hono.client.ServiceInvocationException;
-import org.eclipse.hono.config.ProtocolAdapterProperties;
 import org.eclipse.hono.service.AbstractProtocolAdapterBase;
 import org.eclipse.hono.service.auth.DeviceUser;
 import org.eclipse.hono.service.auth.device.AuthHandler;
@@ -42,6 +42,8 @@ import org.eclipse.hono.service.auth.device.TenantServiceBasedX509Authentication
 import org.eclipse.hono.service.auth.device.UsernamePasswordAuthProvider;
 import org.eclipse.hono.service.auth.device.UsernamePasswordCredentials;
 import org.eclipse.hono.service.auth.device.X509AuthProvider;
+import org.eclipse.hono.service.limiting.ConnectionLimitManager;
+import org.eclipse.hono.service.limiting.MemoryBasedConnectionLimitStrategy;
 import org.eclipse.hono.service.metric.MetricsTags;
 import org.eclipse.hono.service.metric.MetricsTags.Direction;
 import org.eclipse.hono.service.metric.MetricsTags.EndpointType;
@@ -101,7 +103,6 @@ public abstract class AbstractVertxBasedMqttProtocolAdapter<T extends MqttProtoc
     private MqttServer insecureServer;
     private AuthHandler<MqttContext> authHandler;
     private final BiConsumer<CommandSubscription, CommandContext> afterCommandPubAckedConsumer = this::afterCommandPublished;
-    private int connectionLimit = Integer.MAX_VALUE;
 
     /**
      * Sets the authentication handler to use for authenticating devices.
@@ -278,7 +279,9 @@ public abstract class AbstractVertxBasedMqttProtocolAdapter<T extends MqttProtoc
             LOG.warn("authentication of devices turned off");
         }
 
-        connectionLimit = getConfig().isConnectionLimitConfigured() ? checkConnectionLimit() : autoconfigureConnectionLimit();
+        final ConnectionLimitManager connectionLimitManager = Optional.ofNullable(
+                getConnectionLimitManager()).orElse(createConnectionLimitManager());
+        setConnectionLimitManager(connectionLimitManager);
 
         checkPortConfiguration()
             .compose(ok -> {
@@ -291,56 +294,10 @@ public abstract class AbstractVertxBasedMqttProtocolAdapter<T extends MqttProtoc
             }, startFuture);
     }
 
-    private int autoconfigureConnectionLimit() {
-        final long maxMemory = Runtime.getRuntime().maxMemory();
-        final int recommendedLimit = getRecommendedConnectionLimit(maxMemory);
-
-        LOG.info("Connection limit not configured, setting it to {} (based on {} MB memory)",
-                recommendedLimit, maxMemory / 1_000_000);
-
-        return recommendedLimit;
-    }
-
-    private int checkConnectionLimit() {
-        final long maxMemory = Runtime.getRuntime().maxMemory();
-        final int limit = getConfig().getMaxConnections();
-        final int recommendedLimit = getRecommendedConnectionLimit(maxMemory);
-
-        if (limit > recommendedLimit) {
-            LOG.warn("Connection limit ({}) too high: with {} MB memory the limit should be set to max. {}.",
-                    limit, maxMemory / 1_000_000, recommendedLimit);
-        }
-
-        return limit;
-    }
-
-    /**
-     * Returns a recommended limit of concurrent connections for the given maximum amount of memory. An amount of
-     * {@link #MINIMAL_MEMORY} bytes should be reserved for the startup of the adapter and
-     * {@link #MEMORY_PER_CONNECTION} bytes are estimated per connection.
-     * <p>
-     * The value returned is set as the connection limit if no other value is configured in
-     * {@link ProtocolAdapterProperties#setMaxConnections(int)}. If a value is configured and it is above the value
-     * returned here, a warning is printed.
-     * <p>
-     * Subclasses may override this method to implement another calculation.
-     * 
-     * @param maxMemory The maximum amount of memory for the JVM in bytes as returned by {@link Runtime#maxMemory()}.
-     * @return The recommended maximum connection limit between 0 and {@link Integer#MAX_VALUE}.
-     */
-    protected int getRecommendedConnectionLimit(final long maxMemory) {
-
-        final long recommendedLimit = (maxMemory - MINIMAL_MEMORY) / MEMORY_PER_CONNECTION;
-
-        if (recommendedLimit <= 0) {
-            LOG.warn("Not enough memory. It is recommended to provide more than {} MB (currently {} MB).",
-                    MINIMAL_MEMORY / 1_000_000, maxMemory / 1_000_000);
-            return 0;
-        } else if (recommendedLimit > Integer.MAX_VALUE) {
-            return Integer.MAX_VALUE;
-        } else {
-            return (int) recommendedLimit;
-        }
+    private ConnectionLimitManager createConnectionLimitManager() {
+        return new ConnectionLimitManager(
+                new MemoryBasedConnectionLimitStrategy(MINIMAL_MEMORY, MEMORY_PER_CONNECTION),
+                () -> metrics.getNumberOfConnections(), getConfig());
     }
 
     @Override
@@ -396,9 +353,9 @@ public abstract class AbstractVertxBasedMqttProtocolAdapter<T extends MqttProtoc
 
     private Future<Device> handleConnectionRequest(final MqttEndpoint endpoint, final Span currentSpan) {
 
-        if (isConnectionLimitExceeded()) {
-            LOG.debug("Connection limit ({}) exceeded, reject connection request", connectionLimit);
-            currentSpan.log(String.format("connection limit (%d) exceeded", connectionLimit));
+        // the ConnectionLimitManager is null in some unit tests
+        if (getConnectionLimitManager() != null && getConnectionLimitManager().isLimitExceeded()) {
+            currentSpan.log("connection limit exceeded, reject connection request");
             return Future.failedFuture(new MqttConnectionException(
                     MqttConnectReturnCode.CONNECTION_REFUSED_SERVER_UNAVAILABLE));
         }
@@ -408,10 +365,6 @@ public abstract class AbstractVertxBasedMqttProtocolAdapter<T extends MqttProtoc
         } else {
             return handleEndpointConnectionWithoutAuthentication(endpoint);
         }
-    }
-
-    private boolean isConnectionLimitExceeded() {
-        return metrics.getNumberOfConnections() >= connectionLimit;
     }
 
     private void handleConnectionRequestResult(final MqttEndpoint endpoint,