+++ service-base/src/main/java/org/eclipse/hono/service/metric/Metrics.java
@@ -51,7 +51,7 @@ public interface Metrics {
      *
      * @return total number of connections.
      */
-    long getNumberOfConnections();
+    int getNumberOfConnections();
 
     /**
      * Starts a new timer.