+++ service-base/src/test/java/org/eclipse/hono/service/limiting/ConnectionLimitManagerTest.java
@@ -1,0 +1,132 @@
+/*******************************************************************************
+ * Copyright (c) 2019 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+
+package org.eclipse.hono.service.limiting;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+import java.util.function.Supplier;
+
+import org.eclipse.hono.config.ProtocolAdapterProperties;
+import org.junit.Test;
+
+/**
+ * Verifies the behavior of {@link ConnectionLimitManager}.
+ */
+public class ConnectionLimitManagerTest {
+
+    private ConnectionLimitStrategy strategy = mock(ConnectionLimitStrategy.class);
+
+    /**
+     * Verifies that the connection limit is exceeded when it equals the number of connections.
+     */
+    @Test
+    public void testLimitIsEqual() {
+
+        // GIVEN a connection limit of 1
+        final ProtocolAdapterProperties config = new ProtocolAdapterProperties();
+        config.setMaxConnections(1);
+
+        // WHEN there is one connection
+        final Supplier<Integer> currentConnections = () -> 1;
+
+        // THEN the limit is exceeded
+        assertTrue(new ConnectionLimitManager(strategy, currentConnections, config).isLimitExceeded());
+    }
+
+    /**
+     * Verifies that the connection limit is exceeded when it is lower than the number of connections.
+     */
+    @Test
+    public void testLimitIsLower() {
+        // GIVEN a connection limit of 1
+        final ProtocolAdapterProperties config = new ProtocolAdapterProperties();
+        config.setMaxConnections(1);
+
+        // WHEN there are 2 connections
+        final Supplier<Integer> currentConnections = () -> 2;
+
+        // THEN the limit is exceeded
+        assertTrue(new ConnectionLimitManager(strategy, currentConnections, config).isLimitExceeded());
+    }
+
+    /**
+     * Verifies that the connection limit is not exceeded when it is higher than the number of connections.
+     */
+    @Test
+    public void testLimitIsHigher() {
+
+        // GIVEN a connection limit of 2
+        final ProtocolAdapterProperties config = new ProtocolAdapterProperties();
+        config.setMaxConnections(2);
+
+        // WHEN there is one connection
+        final Supplier<Integer> currentConnections = () -> 1;
+
+        // THEN the limit is not exceeded
+        assertFalse(new ConnectionLimitManager(strategy, currentConnections, config).isLimitExceeded());
+    }
+
+    /**
+     * Verifies that the recommended value is used as connection limit if no limit has been configured.
+     */
+    @Test
+    public void testAutoconfigIfNotConfigured() {
+
+        final Supplier<Integer> currentConnections = mock(Supplier.class);
+
+        // GIVEN a ConnectionLimitManager with no limit set and the recommended limit is 2
+        final ProtocolAdapterProperties config = new ProtocolAdapterProperties();
+        when(strategy.getRecommendedLimit()).thenReturn(2);
+        final ConnectionLimitManager connectionLimitManager = new ConnectionLimitManager(strategy, currentConnections,
+                config);
+
+        // WHEN there is one connection
+        when(currentConnections.get()).thenReturn(1);
+        // THEN the limit is not exceeded
+        assertFalse(connectionLimitManager.isLimitExceeded());
+
+        // WHEN there are 2 connections
+        when(currentConnections.get()).thenReturn(2);
+        // THEN the limit is exceeded
+        assertTrue(connectionLimitManager.isLimitExceeded());
+    }
+
+    /**
+     * Verifies that the recommended value is used as connection limit if no config is provided.
+     */
+    @Test
+    public void testAutoconfigIfConfigIsNull() {
+
+        final Supplier<Integer> currentConnections = mock(Supplier.class);
+
+        // GIVEN a ConnectionLimitManager with no config and the recommended limit is 2
+        final ProtocolAdapterProperties config = null;
+        when(strategy.getRecommendedLimit()).thenReturn(2);
+        final ConnectionLimitManager connectionLimitManager = new ConnectionLimitManager(strategy, currentConnections,
+                config);
+
+        // WHEN there is one connection
+        when(currentConnections.get()).thenReturn(1);
+        // THEN the limit is not exceeded
+        assertFalse(connectionLimitManager.isLimitExceeded());
+
+        // WHEN there are 2 connections
+        when(currentConnections.get()).thenReturn(2);
+        // THEN the limit is exceeded
+        assertTrue(connectionLimitManager.isLimitExceeded());
+    }
+}